pyproject.toml
[tool.poetry]
name = "wg-be-exam"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
python-dotenv = "^0.19.0"
uvicorn = "^0.15.0"
SQLAlchemy = "^1.4.25"
psycopg2-binary = "^2.9.1"
databases = "^0.5.2"
asyncpg = "^0.24.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
requests = "^2.26.0"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
